5842: Node Manager only considers nodes busy if they're working.
Previously, Node Manager considered any non-idle node as busy,
including down nodes.  This causes it to boot replacements for nodes
that are marked "down," even if that's because they're still
bootstrapping.  Tighten the busy criteria to avoid booting excess
nodes.

It's easier to make this change now that Node Manager checks that
nodes have a functional Crunch worker state to be considered a
successful bootstrap.  This means that any node that's down later is
in an unexpected state, and we should avoid messing with it.