6661: Install docs refer to arvados-git-httpd package and binary.
The code refers to itself as arv-git-httpd, but the package is named
arvados-git-httpd, and includes the binary under that name as well.
Update the installation documentation to match.

Arguably, we should change the package naming instead.  But that
decision was made longer ago and is more difficult to revert.  I'm
making the easy fix now, and we can consider package changes later if
we like.

Closes #6661.