6663: Fix repository permissions.
Users with is_active==false (except anonymous_user) are not listed by
get_all_permissions.

A repository that is shared with a group is only shared with users who
have permissions on that group, not all users who have permissions on
*any* group.

{User --can_write--> group --can_read--> repo} is correctly listed as
{User --can_read--> repo}, not can_write.

Tests check that the permissions reported by get_all_permissions do
not exceed the API server's interpretation of the current permission
set.