Install Git in shell server install guide.
I don't think we expressly use curl for anything, but reusing the
install template seemed easy enough.  In the future we might want to
expand this to a list of packages that are useful for Arvados work
generally (like maybe jq), but I wanted to keep this commit simple.

No issue #.