6940: Use nginx as WWW_OWNER on Red Hat distributions.
We recommend deploying with Nginx, and this script interacts with
Nginx.  The default Nginx configuration that Phusion ships runs as the
nginx user on these distros.  Use that as the WWW_OWNER.

Ideally we want to move to a more sophisticated place where we
introspect the Web server configuration and use the same server it
does.  But for now, this is an improvement.

No issue #, but discovered when working on #6940.  I cleared this with
Ward.