7435: Node Manager stops trying to shut down delisted cloud nodes.
If the underlying node is gone, trying to destroy it in the cloud will
almost certainly fail.  It's hard to predict what will happen to
related actions like draining the node in SLURM.  Just cancel the
attempt, and trust other systems like SLURM and Crunch to deal with
the disappearance on their own.