7587: Refactor PySDK API tests to use TestCaseWithServers.
History: first this test case used entirely mock responses.  Then we
started running the API server to provide a discovery document.  Then
people added tests that expected to talk to a real test server,
particularly test_empty_list and test_nonempty_list.  These tests
would talk to the API server configured in the user's environment, and
fail if that's not a test API server.

Using TestCaseWithServers fixes the immediate bug in the tests, and
better reflects the current real state of the test case.