6638/7370: Force new builds of Python backports with dependencies.
Even though we've declared these dependencies for a while now, Jenkins
has not published packages with them, because without a new upstream
version, fpm believes that there's no new package to build.  This
resolves that by building a new iteration of the affected packages.

This is less than ideal, because if a new version is released, we'll
automatically package it with iteration 2.  That is not correct, but
it doesn't affect any functionality, and we already have a plan to do
things properly in #6885.  So we'll live with "correct functionality,
gross aesthetics" until then.

Ward approved in conversation.  Refs #6638, #7370.