Install guide crunch-dispatch runit script exports $HOME.
crunch-dispatch needs $HOME set in order to do some path
manipulation.  We set it to $(pwd) in production, which should
normally be the runit service definition directory.

This seems weird and it seems likely that there's a better value, but
until we investigate that and figure that out, documenting what we
have in production is better than letting crunch-dispatch crash.

Reported in IRC.  No issue #.