8082: Ensure consistent column widths in Workbench job banner.
4dcf4e849f242a929ce03b5d529e0e1a63fbaeb0 is the commit that introduced
the column inconsistencies.  This is the smallest patch I could find
to address the issue.  It may be less than ideal because it always
allocates a column to something that may not be displayed, but the
logic to keep track of column widths is already hairy and there are so
many conditions here that it stood to get a lot worse.  Closes #8082.