8371: Rails postrm only fires after removing last RPM.
This prevents removing the package files when the postrm script is
called for the old version of a package after doing an upgrade.

RPM post- scripts are called with $1 as the number of versions of the
package currently installed on the system.  0 means the last version
has been removed.  We only want to remove files in that case.