8203/8690: crunch-job uses `srun --nodes=1`.
Using `srun --nodelist=[one node]` is not sufficient to ensure that
the command runs on a single node; only that it will at least be
allocated the requested node.  This is documented in the srun man
page, and we've observed additional node allocations in these calls.
When srun allocates additional nodes to the work, this can break later
parsing code in crunch-job.

Prefer --nodes, which sets a hard maximum on the number of nodes used.
We're already using this to run individual tasks (-N1), so it's a
known strategy.  Tested live before pushing.  Refs #8203, #8690.