8782: Remove WIFEXITED check from crunch-job reapchildren.
The intent of this check was to avoid reaping children that got
SIGSTOP.  But from the waitpid(2) man page, you must pass specific
flags for waitpid to return those children.  Without those flags,
waitpid will only return the pids of children that have terminated.

Meanwhile, WIFEXITED only returns true if the exit code indicates that
the child terminated normally.  It returns false if the child was
killed by a signal like SIGINT or SIGKILL.  This means children so
killed were not reaped by reapchildren, leading to infinite loops.