8872: Bugfix Node Manager's node search after node create failure.
search_for raises ValueError if the thing isn't found.  create_node
seems to be expecting it to return None instead.  Bring create_node in
line with search_for's documented API.

In order to get the tests to pass, I had to separate out the raw
search code from the caching, and use that in create_node.  Otherwise,
the cloud node from the "node found" test would be cached and returned
in the "node not found" test.