Fix path to docker metadata file
Docker images do not have a `json` file at their root, but rather, a set of
layers, which use a `json` file to manage metadata for the layer. At the root,
the image has a file for metadata named `<hash>.json`. This change instructs
keep to look at that file for the metadata rather than the `json` file.