9463: Unified use of 'replication_desired' param on Collection class at instantiation.
Removed the need to pass the number of copies to be written to keep on save_new() method,
it will be inferred from replication_desired setting or looked up from defaults.

Added functionality to Collection class to keep replication_desired configuration when
loading an already existing collection from API server, with tests validating this new
behaviour.

Corrected some already existing tests to work with this changes.