Adds no_count option to list method
When `no_count` is true, the API server will not query the database
for the count of items and it will not return `items_available` in
the JSON output.

This is expected to offer a major performance improvement when
tables are large and the count is not needed, since postgres
can otherwise take far longer to answer the COUNT() query than
to answer the main query and return the requested list.