We should not hardcode dependencies in our packages for secondary dependencies of our dependencies. If you catch my drift.
pyasn1-modules is a dependency of oauth2client which is a dependency of
google-api-python-client. Which is a dependency of
arvados-python-client.

Pinning a dependency on pyasn1-modules=0.0.5 in arvados-python-client
does not make sense. It also does not appear to be necessary at all. And
it currently breaks the CentOS7 package build.

refs #10713

Revert "8654: Pin pyasn1_modules to version that is compatible with pyasn1==0.1.7."

This reverts commit 436105d3e249ad8dc55fb618d0d2c0d8ed29b395.