12891: Use docker "rm -f" instead of "stop" to cancel container.
"Stop" sends SIGTERM and then (after a timeout) SIGKILL to the main
process in the container. If the contained process catches SIGTERM,
the outcome depends on how much cleanup it manages to do in the
timeout period. We prefer a more predictable SIGKILL outcome.

Arvados-DCO-1.1-Signed-off-by: Tom Clegg <tclegg@veritasgenetics.com>