13064: Check has_ref before checking in_use
Kernel behavior seems to be that if a file is referenced, its parents
remain referenced too.  This means has_ref() exits early when a
collection is not candidate for eviction.

By contrast, in_use() doesn't increment references on parents, so it
requires a full tree walk to determine if a collection is a candidate
for eviction.

Switching the order of the checks retains correctness but greatly
improves real-world performance.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <pamstutz@veritasgenetics.com>