7931: Trash excess replicas when a server has more than one.
Previously a given server was classified only as "having" or "not
having" a replica, regardless of how many it reported. With this
change, once there are already enough distinct replicas on other
servers, excess replicas will be trashed -- even if they are on the
same server as a replica that is being retained.

Arvados-DCO-1.1-Signed-off-by: Tom Clegg <tclegg@veritasgenetics.com>