13993: Support for federated collection requests by uuid
Do remote record retrevial using a salted token, update the block
signatures to "remote" signatures (changing +A to +R and adding the
remote cluster id), and return it to the original client.

Refactor routing and proxying of federated requeststo provide hooks to
support this and future types of requests.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <pamstutz@veritasgenetics.com>