14324: Use logrus in Azure driver.  Fix Sirupsen->sirupsen in imports
"It's in the past been possible to import Logrus as both upper- and
lower-case. Due to the Go package environment, this caused issues in
the community and we needed a standard. Some environments experienced
problems with the upper-case variant, so the lower-case was
decided. Everything using logrus will need to use the lower-case:
github.com/sirupsen/logrus. Any package that isn't, should be
changed."

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <pamstutz@veritasgenetics.com>