13306: Updates ArvCwlExecutor to properly convert objects to JSON in unicode
json.dumps produces a str object in python 2 (bytestring) and a str object in python 3 (unicode), resulting in incompatibility when writing to the file (expects unicode). In order to prevent using py2 and py3 specific code, the solution is to call encode, which will force the string to be bytes in both languages, then decode to unicode from there. 

Arvados-DCO-1.1-Signed-off-by:  Eric Biagiotti <ebiagiotti@veritasgenetics.com>