13306: Removing unneeded try/except statements in the acr submit tests
Removed the try/except statements and moved the asserts checking if the call to arvados_cwl.main exited properly to the bottom of the tests.  In other words, the tests should be using specific call asserts to establish if it ran correctly and checking the exited value should be a last resort to report an unhandled case.

Arvados-DCO-1.1-Signed-off-by:  Eric Biagiotti <ebiagiotti@veritasgenetics.com>