14718: Issue salted token when passing 'remote' param to login.
When logging in with a 'remote' parameter, it's encoded into the return_to
param passed to the SSO server so that it's received back after the user
is authenticated.
When the omniauth callback receives a remote parameter inside the return_to
url, it returns the salted version of the client token.

Arvados-DCO-1.1-Signed-off-by: Lucas Di Pentima <ldipentima@veritasgenetics.com>