14885: Removes as_naive datetime parsing distinction
parse_datetime_as_naive should only be used in instances where we want to parse a string that we know contains timezone info AND  we do not want to include it in the resulting datetime object. Otherwise, parse_datetime is preferred for parsing strings with or without timezone info.

Arvados-DCO-1.1-Signed-off-by: Eric Biagiotti <ebiagiotti@veritasgenetics.com>