14873: Upgrades 'pg' gem & solves an issue related to a new internal table.
Rails 5 includes an undocumented internal table called 'ar_internal_metadata'
used to record the environment that the current system is on, so that it can
avoid accidental database wipes when in production.
Our code iterates over the tables so this one needs to be skipped.

Arvados-DCO-1.1-Signed-off-by: Lucas Di Pentima <ldipentima@veritasgenetics.com>