14873: Custom JSON attributes that default to [] or {} when nil.
Previously we used the 'serialize' call to declare a JSONB column as a Hash
or Array and this had the side effect to default to an empty hash or array
when assigning nil (whether was reading from the database or passed as a param).
JSONB columns shouldn't be now declared as serialized because Rails would serialize
their contents twice, but not declaring them makes us lose their side effect.

Arvados-DCO-1.1-Signed-off-by: Lucas Di Pentima <ldipentima@veritasgenetics.com>