14717: Fix fallback behavior for component config vs main config
For backwards compatability, we need to be able to start with only a
component config and not the main config, and it is a fatal error if
it doesn't exist.

However, if there is a main config, and then it is okay if the
component config doesn't exist.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <pamstutz@veritasgenetics.com>