15521: Fix truncated index response.
The client (keep-balance) looks for a blank line at the end of the
index response as an assurance the response is not truncated.  If
headers have already been sent, http.Error("") writes a blank line --
so we were inadvertently assuring the client that truncated responses
were not truncated.

Arvados-DCO-1.1-Signed-off-by: Tom Clegg <tclegg@veritasgenetics.com>