15736: When validating sessions, save the config
All clusters in 'sessions' (whether part of the federation or not)
where we are able to successfully fetch either the new exported public
config or the old discovery document will get added to
auth.remoteHostsConfig.  This means links to other clusters
works consistently (sort of, users may still need to log in.)

Also, it now updates the title bar with the cluster id.