15803: Remove the 'email' can_login permission created for new users
On setup, it used to create a 'permission' link with user email
address 'tail' pointing to a user uuid 'head'.  The identity_url was
set on properties.

However, since #15558 this link is no longer read by anything in the
API server (except the test suite).  Since it is confusing and a
potential source of error, this gets rid of it.

Also fixes some typos.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <pamstutz@veritasgenetics.com>