16139: Fix secondaryFile errors when running --submit --no-wait
When submitting, we want to preserve the original CWL version of the
document.

However, when it creates a RunnerContainer (a cwltool.Process) it
examines the input interface and expects it to be in the 1.1 data
model.  But if we preserve/reload the original document, it is still
in the 1.0 data model.  This causes

The code was deliberately overriding the CWL version in metadata to
make this work.  This results in the problem reported on this ticket.

The fix is to maintain both the updated and preserved documents, and
use them appropriately where they are expected.

refs #16139

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>