Remove Gemfile.lock from login-sync
no issue #

To be consistent with sdk/ruby sdk/cli behavior.

There's a circular dependency where the version number of the gem is
embedded in Gemfile.lock, but the version number is derived from the
git commit timestamp, so when it updates Gemfile.lock it creates a new
version, which requires a updating Gemfile.lock, and so on.

However, unlike the Ruby on Rails deb/rpm packages, when we create
gems (dev or production) it does not include Gemfile.lock, so there's
no benefit to having it checked in.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>