16118: Adds test checking writable/readonly collection UI changes.
WIP: There's deactivated code that shows a snackbar whenever a service request
returns an error, no matter if the error is handled somewhere up in the stack.
I think that isn't a good approach, also it prevents the 'readonly' case to
work because the snackbar appears in fron of a menu button and cannot be
clicked.

Arvados-DCO-1.1-Signed-off-by: Lucas Di Pentima <lucas@di-pentima.com.ar>