15610: Uses a virtualized list to show the collection's file tree. (WIP)
This greatly improves rendering times when showing collections with many
files on a directory.
This is a POC: the whole tree is rendered expanded to show that it doesn't
affect render times, it still needs lots of tweaking to offer the same
behavior as before.

Arvados-DCO-1.1-Signed-off-by: Lucas Di Pentima <lucas@di-pentima.com.ar>