16470: Avoids crashing when running db:create in development mode.
The 'preload_all_models' initializer is now run even when db:create is
called, so it avoids the database to be created on arvbox, for example,
because it attempts to access the database before being created.

This initializer was added 7 years ago and it now seems to not be needed
anymore, so I think it's the cleaner way to resolve this issue.

See: https://github.com/rails/rails/issues/32870

Arvados-DCO-1.1-Signed-off-by: Lucas Di Pentima <lucas@di-pentima.com.ar>