16826: Enhances teardown config check.
It's not enough to check that .keys() returns a list of Symbols, because
ActiveRecord::OrderedOptions behaves somewhat like HashWithIndifferentAccess
so values can need to be accessed by symbol or string. So, we now check that
every hash-like config section is an OrderedOptions object and also returns
its keys as symbols.

Arvados-DCO-1.1-Signed-off-by: Lucas Di Pentima <lucas@di-pentima.com.ar>