container requests in the new codepath for controller
Controller's "new" codepath is a step towards sunsetting railsapi. This
commit is the initial work to get Container Requests in that path

The reason for deleting some tests is intentional. This is part of the migration
from
   apps/workbench/test/controllers/container_requests_controller_test.rb
to
   lib/controller/integration_test.go

This single-commit is the result of collaboration with Lucas Di Pentima
and Tom Clegg in the branch 17014-controller-cr

Arvados-DCO-1.1-Signed-off-by: Nico Cesar <nico@curii.com>